Home page component added
